generateSettings() moved to settings.ts
